LayoutTests/fast/dom/Window/open-window-min-size.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
function test() {
    var expectedMsgs = 2;
    var msgsReceived = 0;
    window.addEventListener('message', function(e) {
        msgsReceived++;
        shouldBeGreaterThanOrEqual(e.data["width"].toString(), "100");
        shouldBeGreaterThanOrEqual(e.data["height"].toString(), "100");
        if (msgsReceived === expectedMsgs)
            testRunner.notifyDone();
    }, false);

    window.open("resources/print-close.html", "", "width=10, height=10");
    window.showModalDialog("resources/print-close.html", "", "dialogWidth:10; dialogHeight:10");
}

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
    testRunner.setCanOpenWindows();
    testRunner.setCloseRemainingWindowsWhenComplete(true);
}

</script>
</head>
<body onload="test()">
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
